+++ ReleaseNotes.md
@@ -1,5 +1,7 @@
 # 3.2.2
 
+- Fixed a bug with WPF custom control libraries (#178, #179, thanks @markrendle)
+
 # 3.2.1
 
 - Sets SolutionFilePath in the Workspace Solution when it's available (#177, thanks @markrendle)